backend/requirements.txt
asgiref==3.7.2
Brotli==1.1.0
click==8.1.7
colorama==0.4.6
dj-database-url==2.1.0
Django==5.0.3
django-cors-headers==4.3.1
django-filter==23.5
djangorestframework==3.14.0
gunicorn==21.2.0
h11==0.14.0
Markdown==3.5.2
packaging==23.2
psycopg2-binary==2.9.9
pytz==2024.1
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
uvicorn==0.27.1
whitenoise==6.6.0
